Consistent sentinels for event-logging IDs, encounter types, symptoms (#16)
* Add conda environment.yml

Reproducible conda environment for building and running JUNE2. Includes cxx-compiler (GCC on Linux), cmake, hdf5, yaml-cpp, openmpi, Python test deps (h5py, numpy), and optional deps metis and gperftools.

* Consistent sentinels for event-logging IDs, encounter types, symptoms

Event-logging sentinel/fallback choices had drifted: bare -1/255 literals
repeated with no named constant, and infector_symptom_id used 0 both as
the real "recovered" registry id and as Disease::getSymptomId's own
error fallback, so seed-infection rows (no real infector) were
indistinguishable from a recovered infector.

- Add kInvalidPersonId/kInvalidVenueId (-1), kUnknownEncounterTypeId (255),
  kNoSymptomId (255) in core/types.h; replace bare literals across
  PendingInfection, PersonLocation, config.h encounter-type fields.
- Narrow infector_symptom_id/old_symptom_id/new_symptom_id uint16_t -> uint8_t
  end to end (event structs, HDF5 CompType registrations, logger API),
  freeing symptom id 0 to mean only "recovered". Disease's internal
  uint16_t representation is untouched; narrowing happens only at the
  logInfection/logSymptomChange call boundary via explicit static_cast.
- Standardise "nothing to report" string fallbacks in lookup tables:
  "unknown" -> "could not resolve" (value existed but couldn't be
  resolved), "" -> "not applicable" (nothing to report).

Verified: full build passes, test_integration and test_simulator both
pass, and a standalone EventLogger spot-check confirms the HDF5 output
now has uint8 symptom-id columns with 255 (not 0) for seed infections.

* Fix leftover uint16_t/default-value gaps from event-logging sentinel narrowing

Commit dbccfd1 narrowed infector_symptom_id (and related event-logging
fields) to uint8_t and introduced kUnknownEncounterTypeId/kNoSymptomId
sentinels, but two call sites weren't updated to match:

- interaction_manager_partial_presence.cpp: unpacking MPI-packed
  candidate data cast infector_symptom_id to uint16_t, then relied on an
  implicit second narrowing to fit PendingInfection's now-uint8_t field.
  Cast directly to uint8_t instead, matching the field type and the
  explicit style used at the sibling call site in
  interaction_manager_venue.cpp.

- infection_seed.cpp: logInfection call for seed infections omitted the
  encounter-type and symptom-id sentinel arguments added by dbccfd1
  (no infector exists for seeded infections), leaving them at
  compiler/default-initialised values instead of the intended
  kUnknownEncounterTypeId/kNoSymptomId sentinels.

* Finish sentinel-consistency cleanup from dbccfd1

Leftover raw `255` comparisons (interaction_manager.cpp,
interaction_manager_partial_presence.cpp, activity_manager_selection.cpp)
and duplicated fallback strings ("could not resolve", "not applicable" in
event_writer_lookups.cpp, event_merger_lookups.cpp) sat outside the single
point of definition dbccfd1 meant to establish, reintroducing the same
drift risk it fixed elsewhere. Replace with kUnknownEncounterTypeId and
new kCouldNotResolve/kNotApplicable constants (event_types.h).

* Rename kUnknownEncounterTypeId to kDefaultEncounterTypeId; split kUnknownVenueTypeId

encounter_type_id is 255 for every ordinary, non-coordinated encounter
(the vast majority), only getting a real encounter_type_names index for
coordinated encounters. "Unknown" read as an error/undefined state, but
this is the expected default for the staple case, so rename to
kDefaultEncounterTypeId (still 255, purely a naming fix, no behaviour
change).

Considered promoting "uncoordinated" to a real registry entry instead,
but rejected: interaction_manager.cpp and interaction_manager_venue_bins.cpp
use encounter_type_id < encounter_type_names.size() as an implicit
"is this coordinated" test, and the partial-presence v1 invariant relies
on the sentinel staying out of the registry's range, so a real in-range
default id would corrupt encounter stats.

Also split kUnknownEncounterTypeId's accidental reuse in
activity_manager_selection.cpp (testing a venue-type lookup miss from
getVenueTypeId, an unrelated concept) into its own kUnknownVenueTypeId
constant.

* Finish sentinel-constant rollout in interaction_manager

Three venue_type_id defaults in the interaction_manager module still used
the raw literal 255, left over from the event-logging sentinel-consistency
cleanup (dbccfd1..ac7ad42) that introduced kUnknownVenueTypeId. Swap them
in for consistency: ParentAggregate::parent_venue_type_id,
processVenueTransmissions' local venue_type_id, and
resolveVenueTypeAndMatrix's output default.

* Fix MPI wire-format desync in PendingInfection pack/unpack

infector_symptom_id narrowed uint16_t->uint8_t (dbccfd1..ac7ad42) shrank
the packed record by 1 byte, but INFECTION_SIZE stayed a hardcoded 25 and
unpackAndApplyIncoming's local still declared uint16_t. Alltoallv slots
stayed 25 bytes/record while packing wrote 24, so any batch of >1
PendingInfection to the same destination rank desynced from the second
record onwards, corrupting person_id/infector_id/venue etc with no crash.

Fix: derive INFECTION_SIZE/PROPOSAL_WIRE_SIZE/REPLY_WIRE_SIZE from
sizeof(decltype(field)) of the actual struct members instead of literals,
and make unpackAndApplyIncoming's locals decltype-derived too, so a future
struct field-type change can't silently drift pack/unpack apart again.
EncounterReply's status_byte stays a manual uint8_t (deliberate enum
transcode, not drift).

Add regression test: two PendingInfection records routed to the same
destination rank in one round, asserting both round-trip intact.

* Fix resolveInfectorSymptomId no-infector sentinel mismatch

Was returning raw 0 for "no infector", ambiguous with a real symptom
stage 0 from getCurrentSymptomId. infection_seed.cpp already logs
kNoSymptomId (255) for the equivalent seed-infection case; align
resolveInfectorSymptomId's two no-infector paths (infector_id < 0,
visitor not found) to the same sentinel so venue/partial-presence
infections aren't misread by downstream tooling.

* Shorten kCouldNotResolve to fit PersonRecord::sex[16]

"could not resolve" (18 chars) truncated to "could not resol" when
strncpy'd into the 16-byte sex field, silently corrupting HDF5 output
for Sex::UNKNOWN. Shortened sentinel to "unresolved" (10 chars) so it
fits all fixed-size buffer call sites without truncation.

* Finish sentinel rollout for cached_virtual_venue_type_id

Sibling field cached_encounter_type_id was converted to kDefaultEncounterTypeId
in this branch's rollout, but cached_virtual_venue_type_id was left as bare
255 despite serving the same "not yet resolved" role. Use kUnknownVenueTypeId
instead. Field stays int, not uint8_t: it's compared against
EncounterProposal::venue_type_id and populated from
ContactMatrices::matrix_name_to_id, both int, so int is the correct type for
this subsystem.

* Guard uint16_t->uint8_t symptom-id narrowing at Disease load

infector_symptom_id/old_symptom_id/new_symptom_id were narrowed to
uint8_t for event logging and cross-rank transmission, but symptom-tag
ids are assigned unbounded from YAML config with no cap. >255 tags
would silently wrap the id mod 256 at every downstream call site.
Throw at load time instead if symptom_tags count reaches kNoSymptomId
(255, the "not applicable" sentinel).

* Apply clang-format to fix CI lint failure

Reflow arg lists in infectPerson, filterAndSortActiveLocations,
resolvePartialPresenceInfections, buildPersonRecords to satisfy
80-col clang-format check; no semantic change.

---------

+++ include/parallel/domain_communicator.h
@@ -101,7 +101,7 @@ class DomainCommunicator {
   // disease loaded).
   std::optional<PendingInfection> applyOnePendingInfection(
       PersonId pid, PersonId infector_id, double t, uint8_t v_type,
-      uint8_t enc_type_id, VenueId v_id, uint16_t infector_symptom_id,
+      uint8_t enc_type_id, VenueId v_id, uint8_t infector_symptom_id,
       uint8_t transmission_mode_index);
 
   WorldState& world_;
